+++ pkg/manifest/local_server.go
@@ -13,7 +13,7 @@ type LocalServer struct {
 	ConfigStores   map[string]LocalConfigStore `toml:"config_stores,omitempty"`
 	KVStores       LocalKVStoreMap             `toml:"kv_stores,omitempty"`
 	SecretStores   LocalSecretStoreMap         `toml:"secret_stores,omitempty"`
-	Pushpin        LocalPushpinMap             `toml:"pushpin,omitempty"`
+	Pushpin        *LocalPushpinMap            `toml:"pushpin,omitempty"`
 	ViceroyVersion string                      `toml:"viceroy_version,omitempty"`
 }
 
@@ -196,10 +196,10 @@ func (m *LocalSecretStoreMap) UnmarshalTOML(v any) error {
 // LocalPushpinMap represents configuration of a local instance of Pushpin,
 // used for local experimentation and testing of handoff_fanout.
 type LocalPushpinMap struct {
-	EnablePushpin      bool   `toml:"enable,omitempty"`
-	PushpinPath        string `toml:"pushpin_path,omitempty"`
-	PushpinProxyPort   uint16 `toml:"proxy_port,omitempty"`
-	PushpinPublishPort uint16 `toml:"publish_port,omitempty"`
+	EnablePushpin      *bool   `toml:"enable,omitempty"`
+	PushpinPath        *string `toml:"pushpin_path,omitempty"`
+	PushpinProxyPort   *uint16 `toml:"proxy_port,omitempty"`
+	PushpinPublishPort *uint16 `toml:"publish_port,omitempty"`
 }
 
 func decodeTOMLMap(m map[string]any, out any) error {